VerificationEngine.kt -> FileX

+++ app/src/main/java/balti/migrate/backupEngines/engines/VerificationEngine.kt
@@ -2,12 +2,15 @@ package balti.migrate.backupEngines.engines
 
 import android.content.pm.PackageInfo
 import android.util.Log
+import balti.filex.FileX
+import balti.migrate.AppInstance.Companion.CACHE_DIR
 import balti.migrate.R
 import balti.migrate.backupEngines.BackupServiceKotlin
 import balti.migrate.backupEngines.ParentBackupClass
 import balti.migrate.backupEngines.containers.BackupIntentData
 import balti.migrate.backupEngines.utils.BackupUtils
 import balti.migrate.extraBackupsActivity.apps.containers.AppPacket
+import balti.migrate.utilities.CommonToolsKotlin
 import balti.migrate.utilities.CommonToolsKotlin.Companion.DEBUG_TAG
 import balti.migrate.utilities.CommonToolsKotlin.Companion.ERR_CORRECTION_SHELL
 import balti.migrate.utilities.CommonToolsKotlin.Companion.ERR_CORRECTION_SUPPRESSED
@@ -28,10 +31,8 @@ import balti.module.baltitoolbox.functions.FileHandlers.getDirLength
 import balti.module.baltitoolbox.functions.Misc.getPercentage
 import balti.module.baltitoolbox.functions.Misc.tryIt
 import balti.module.baltitoolbox.functions.SharedPrefs.getPrefBoolean
-import java.io.File
 import java.io.BufferedReader
 import java.io.BufferedWriter
-import java.io.FileWriter
 import java.io.InputStreamReader
 import java.io.OutputStreamWriter
 
@@ -60,7 +61,7 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
         allErrors.add(err)
     }
 
-    private fun getDataCorrectionCommand(pi: PackageInfo, expectedDataFile: File): String{
+    private fun getDataCorrectionCommand(pi: PackageInfo, expectedDataFile: FileX): String{
 
         val fullDataPath = pi.applicationInfo.dataDir
         val actualDataName = fullDataPath.substring(fullDataPath.lastIndexOf('/') + 1)
@@ -69,13 +70,13 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
         return "if [ -e $fullDataPath ]; then\n" +
                 "   cd $dataPathParent\n" +
                 "   echo \"Copy data: ${expectedDataFile.name}\"\n" +
-                "   $busyboxBinaryPath tar -vczpf ${expectedDataFile.absolutePath} $actualDataName\n" +
+                "   $busyboxBinaryPath tar -vczpf ${expectedDataFile.canonicalPath} $actualDataName\n" +
                 "fi\n\n"
     }
 
     private fun getDataCorrectionCommand(dataName: String): String {
         val packageName = dataName.let { it.substring(0, it.lastIndexOf(".tar.gz")) }
-        val expectedDataFile = File(actualDestination, dataName)
+        val expectedDataFile = FileX.new(actualDestination, dataName)
 
         return getDataCorrectionCommand(pm.getPackageInfo(packageName, 0), expectedDataFile)
     }
@@ -101,22 +102,22 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
 
                 broadcastProgress(appName, "verifying: $appName", false)
 
-                val expectedAppDir = File(actualDestination, "$packageName.app")
-                val expectedApkFile = File("$actualDestination/$packageName.app", "$packageName.apk")
-                val expectedDataFile = File("$actualDestination/$packageName.tar.gz")
-                val expectedPermFile = File(actualDestination, "$packageName.perm")
+                val expectedAppDir = FileX.new(actualDestination, "$packageName.app")
+                val expectedApkFile = FileX.new("$actualDestination/$packageName.app", "$packageName.apk")
+                val expectedDataFile = FileX.new("$actualDestination/$packageName.tar.gz")
+                val expectedPermFile = FileX.new(actualDestination, "$packageName.perm")
 
                 if (getPrefBoolean(PREF_NEW_ICON_METHOD, true) ) {
-                    if (!File(actualDestination, "$packageName.png").exists()) allRecovery.add("$MIGRATE_STATUS:icon_new:$packageName")
+                    if (!FileX.new(actualDestination, "$packageName.png").exists()) allRecovery.add("$MIGRATE_STATUS:icon_new:$packageName")
                 }
                 else {
-                    if (!File(actualDestination, "$packageName.icon").exists()) allRecovery.add("$MIGRATE_STATUS:icon_old:$packageName")
+                    if (!FileX.new(actualDestination, "$packageName.icon").exists()) allRecovery.add("$MIGRATE_STATUS:icon_old:$packageName")
                 }
 
                 if (packet.APP) {
 
                     val existsAppDir = expectedAppDir.exists()
-                    val sizeAppDir = getDirLength(expectedAppDir.absolutePath)
+                    val sizeAppDir = expectedAppDir.getDirLength()
                     val existsApk = expectedApkFile.exists()
                     val sizeApk = expectedApkFile.length()
 
@@ -128,8 +129,8 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
 
                         apkName = commonTools.applyNamingCorrectionForShell(apkName)
 
-                        expectedAppDir.absolutePath.let {
-                            /*if (CommonToolsKotlin.isDeletable(expectedAppDir)) {
+                        expectedAppDir.canonicalPath.let {
+                            if (CommonToolsKotlin.isDeletable(expectedAppDir)) {
                                 allRecovery.add(
                                         "echo \"Copy apk(s): $packageName\"\n" +
                                                 "rm -rf $it 2> /dev/null\n" +
@@ -138,7 +139,7 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
                                                 "cp *.apk $it/\n\n" +
                                                 "mv $it/$apkName $it/${pi.packageName}.apk\n"
                                 )
-                            }*/
+                            }
                         }
 
                         broadcastProgress(appName, "$packageName : appDir $existsAppDir, $sizeAppDir : apk $existsApk, $sizeApk", false)
@@ -194,41 +195,45 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
 
             resetBroadcast(false, title, EXTRA_PROGRESS_TYPE_VERIFYING)
 
-            val tarCheckScript = File(engineContext.filesDir.absolutePath, "$FILE_PREFIX_TAR_CHECK.sh")
-            val scriptWriter = BufferedWriter(FileWriter(tarCheckScript))
+            val tarCheckScript = FileX.new(CACHE_DIR, "$FILE_PREFIX_TAR_CHECK.sh", true)
+            //val scriptWriter = BufferedWriter(FileWriter(tarCheckScript))
 
-            scriptWriter.write("#!sbin/sh\n\n")
-            scriptWriter.write("echo \" \"\n")
-            scriptWriter.write("sleep 1\n")
-            scriptWriter.write("echo \"--- TAR CHECK PID: $$\"\n")
-            scriptWriter.write("cp ${tarCheckScript.absolutePath} ${engineContext.externalCacheDir}/\n")
+            tarCheckScript.startWriting(object : FileX.Writer(){
+                override fun writeLines() {
 
-            scriptWriter.write(
-                    "checkData(){\n" +
-                            "\n" +
-                            "    echo \"--- TAR_GZ \$1 ---\"\n" +
-                            "    err=\"\$(${busyboxBinaryPath} gzip -t \"${actualDestination}/\$1\" 2>&1)\"\n" +
-                            "    if [[ ! -z \"\$err\" ]]; then\n" +
-                            "        echo \"--- ERROR:\$1:\$err ---\"\n" +
-                            "    else\n" +
-                            "        echo \"--- OK:\$1 ---\"\n" +
-                            "    fi\n" +
-                            "\n" +
-                            "}\n"
-            )
+                    write("#!sbin/sh\n\n")
+                    write("echo \" \"\n")
+                    write("sleep 1\n")
+                    write("echo \"--- TAR CHECK PID: $$\"\n")
+                    write("cp ${tarCheckScript.absolutePath} ${CACHE_DIR}/\n")
 
-            for (i in appList.indices) {
+                    writeLine(
+                            "checkData(){\n" +
+                                    "\n" +
+                                    "    echo \"--- TAR_GZ \$1 ---\"\n" +
+                                    "    err=\"\$(${busyboxBinaryPath} gzip -t \"${actualDestination}/\$1\" 2>&1)\"\n" +
+                                    "    if [[ ! -z \"\$err\" ]]; then\n" +
+                                    "        echo \"--- ERROR:\$1:\$err ---\"\n" +
+                                    "    else\n" +
+                                    "        echo \"--- OK:\$1 ---\"\n" +
+                                    "    fi\n" +
+                                    "\n" +
+                                    "}\n"
+                    )
 
-                val packet = appList[i]
-                val expectedDataFile = File("$actualDestination/${packet.PACKAGE_INFO.packageName}.tar.gz")
+                    for (i in appList.indices) {
 
-                if (packet.DATA && expectedDataFile.exists() && expectedDataFile.length() != 0L) {
-                    scriptWriter.write("checkData ${expectedDataFile.name}\n")
+                        val packet = appList[i]
+                        val expectedDataFile = FileX.new("$actualDestination/${packet.PACKAGE_INFO.packageName}.tar.gz")
+
+                        if (packet.DATA && expectedDataFile.exists() && expectedDataFile.length() != 0L) {
+                            write("checkData ${expectedDataFile.name}\n")
+                        }
+                    }
+
+                    writeLine("echo \"--- Tar checks complete ---\"\n")
                 }
-            }
-
-            scriptWriter.write("echo \"--- Tar checks complete ---\"\n")
-            scriptWriter.close()
+            })
 
             suProcess = Runtime.getRuntime().exec("su")
             suProcess?.let {
@@ -316,47 +321,50 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
         resetBroadcast(false, title, EXTRA_PROGRESS_TYPE_CORRECTING)
 
         try {
-            val retryScript = File(engineContext.filesDir.absolutePath, "$FILE_PREFIX_RETRY_SCRIPT.sh")
-            val scriptWriter = BufferedWriter(FileWriter(retryScript))
+            val retryScript = FileX.new(CACHE_DIR, "$FILE_PREFIX_RETRY_SCRIPT.sh", true)
+            retryScript.startWriting(object : FileX.Writer(){
+                override fun writeLines() {
+                    write("#!sbin/sh\n\n")
+                    write("echo \" \"\n")
+                    write("sleep 1\n")
+                    write("echo \"--- RECOVERY PID: $$\"\n")
+                    write("cp ${retryScript.absolutePath} ${CACHE_DIR}/\n")
 
-            scriptWriter.write("#!sbin/sh\n\n")
-            scriptWriter.write("echo \" \"\n")
-            scriptWriter.write("sleep 1\n")
-            scriptWriter.write("echo \"--- RECOVERY PID: $$\"\n")
-            scriptWriter.write("cp ${retryScript.absolutePath} ${engineContext.externalCacheDir}/\n")
+                    for (i in 0 until defects.size){
+                        if (BackupServiceKotlin.cancelAll) break
 
-            for (i in 0 until defects.size){
-                if (BackupServiceKotlin.cancelAll) break
-
-                val defect = defects[i]
-                if (defect.startsWith(MIGRATE_STATUS)){
-                    val parts = defect.split(":")
-                    if (parts.size == 3){
-                        when (parts[1]){
-                            "icon_new" -> {
-                                val pi = pm.getPackageInfo(parts[2], 0)
-                                backupUtils.makeNewIconFile(parts[2], iconTools.getBitmap(pi, pm), actualDestination)
-                            }
-                            "icon_old" -> {
-                                val pi = pm.getPackageInfo(parts[2], 0)
-                                backupUtils.makeStringIconFile(parts[2], iconTools.getIconString(pi, pm), actualDestination)
-                            }
-                            "perm" -> {
-                                backupUtils.makePermissionFile(parts[2], actualDestination, pm)
+                        val defect = defects[i]
+                        if (defect.startsWith(MIGRATE_STATUS)){
+                            val parts = defect.split(":")
+                            if (parts.size == 3){
+                                when (parts[1]){
+                                    "icon_new" -> {
+                                        val pi = pm.getPackageInfo(parts[2], 0)
+                                        backupUtils.makeNewIconFile(parts[2], iconTools.getBitmap(pi, pm), actualDestination)
+                                    }
+                                    "icon_old" -> {
+                                        val pi = pm.getPackageInfo(parts[2], 0)
+                                        backupUtils.makeStringIconFile(parts[2], iconTools.getIconString(pi, pm), actualDestination)
+                                    }
+                                    "perm" -> {
+                                        backupUtils.makePermissionFile(parts[2], actualDestination, pm)
+                                    }
+                                }
+                                write("echo \"--- DEFECT: ${i + 1}\"\n")
+                                write("echo \"$defect\"\n")
                             }
                         }
-                        scriptWriter.write("echo \"--- DEFECT: ${i + 1}\"\n")
-                        scriptWriter.write("echo \"$defect\"\n")
+                        else {
+                            write("echo \"--- DEFECT: ${i + 1}\"\n")
+                            write(defect)
+                        }
                     }
-                }
-                else {
-                    scriptWriter.write("echo \"--- DEFECT: ${i + 1}\"\n")
-                    scriptWriter.write(defect)
-                }
-            }
 
-            scriptWriter.write("echo \"--- Retry complete ---\"\n")
-            scriptWriter.close()
+                    write("echo \"--- Retry complete ---\"\n")
+                }
+            })
+
+
 
             suProcess = Runtime.getRuntime().exec("su")
             suProcess?.let {
@@ -421,7 +429,7 @@ class VerificationEngine(private val jobcode: Int, private val bd: BackupIntentD
         }
     }
 
-    override suspend fun doInBackground(arg: Any?): Any? {
+    override suspend fun doInBackground(arg: Any?): Any {
 
         val defects = verifyBackups()
         if (!BackupServiceKotlin.cancelAll && defects != null && defects.size != 0)
